+++ src/components/User/SystemAssets/components/AssetManager.vue
@@ -237,6 +237,16 @@
 
       <template v-if="isShownAdvanced">
         <div class="app__form-row">
+          <tick-field
+            title="Use Coinpayments service for deposition"
+            class="app__form-field"
+            v-model="asset.creatorDetails.isCoinpayments"
+            label="Use Coinpayments"
+            :disabled="isPending"
+          />
+        </div>
+
+        <div class="app__form-row">
           <input-field
             class="app__form-field app__form-field--halved"
             type="number"
@@ -310,6 +320,7 @@ export default {
           name: '',
           logo: {},
           terms: {},
+          isCoinpayments: false,
           externalSystemType: ''
         }
       },
@@ -392,6 +403,7 @@ export default {
             creatorDetails: {
               name: this.asset.creatorDetails.name,
               isFiat: this.asset.creatorDetails.isFiat,
+              is_coinpayments: this.asset.creatorDetails.isCoinpayments,
               external_system_type: this.asset.creatorDetails.externalSystemType,
               logo,
               terms
@@ -412,6 +424,7 @@ export default {
               name: this.asset.creatorDetails.name,
               isFiat: this.asset.creatorDetails.isFiat,
               external_system_type: this.asset.creatorDetails.externalSystemType,
+              is_coinpayments: this.asset.creatorDetails.isCoinpayments,
               logo: {
                 key: this.asset.creatorDetails.logo.key,
                 type: this.asset.creatorDetails.logo.type